requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.19.0
azure-cosmos==4.7.0
azure-kusto-data==4.4.1
azure-kusto-ingest==4.4.1
pandas==2.2.2
